requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==1.0.3
python-multipart==0.0.6
Jinja2==3.1.2
aiofiles==23.2.1
websockets==11.0.3
passlib[bcrypt]==1.7.4
python-magic==0.4.27